revise

+++ pykt/datasets/lpkt_dataloader.py
@@ -250,7 +250,8 @@ class LPKTDataset(Dataset):
                 dqtest["orirow"].append([int(_) for _ in row["orirow"].split(",")])
         for key in dori:
             if key not in ["rseqs"]:#in ["smasks", "tseqs"]:
-                dori[key] = LongTensor(dori[key])
+                dori[key] = FloatTensor(dori[key])
+                # dori[key] = LongTensor(dori[key])
             else:
                 dori[key] = FloatTensor(dori[key])
